requirements.txt
gym>=0.26.2
imageio>=2.23.0
ipdb>=0.13.11
minatar==1.0.13
numpy>=1.23.5
Pillow>=9.4.0
scikit_learn>=1.2.1
scipy>=1.10.0
setuptools>=65.5.0
typing_extensions>=4.4.0
imageio[ffmpeg]>=2.26.0
imageio[pyav]>=2.26.0
wandb>=0.14.0
opencv-python>=4.7.0.72
torch>=2.0.0
scikit-video>=1.1.11